Only test for the request body in the existing "global-type" test
This renames the existing "global-type" testcase to "global-type-body" as
the current test only ensures that a global type can be passed as body.
This makes the "User" schema in the global-type-body testcase obsolete,
which is why it's being removed.
A new test which ensures that a global-type can be used as return type
will be added separately.
