+++ raml-client-generator-core/src/test/resources/global-type-body/api.raml
@@ -6,26 +6,6 @@ documentation:
   - title : Foo
     content : A String
 schemas:
-    - user: |
-        {
-          "$schema": "http://json-schema.org/draft-04/schema#",
-          "id": "http://jsonschema.net",
-          "type": "object",
-          "properties": {
-            "username": {
-              "id": "http://jsonschema.net/username",
-              "type": "string"
-            },
-            "password": {
-              "id": "http://jsonschema.net/password",
-              "type": "string"
-            }
-          },
-          "required": [
-            "username",
-            "password"
-          ]
-        }
     - auth: |
         {
           "$schema": "http://json-schema.org/draft-04/schema#",
@@ -53,7 +33,6 @@ schemas:
         200:
           body:
             application/json:
-              schema: user
       body:
         application/json:
           schema: auth
