+++ raml-client-generator-core/src/test/resources/global-type-body/output/global-type-body/api/FooClient.java
@@ -1,9 +1,9 @@
 
-package global-type.api;
+package global-type-body.api;
 
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
-import global-type.resource.cs.Cs;
+import global-type-body.resource.cs.Cs;
 
 
 /**
