+++ raml-client-generator-core/src/test/resources/global-type-body/output/global-type-body/exceptions/FooException.java
@@ -1,5 +1,5 @@
 
-package global-type.exceptions;
+package global-type-body.exceptions;
 
 
 public class FooException
