+++ raml-client-generator-core/src/test/resources/global-type-body/output/global-type-body/resource/cs/Cs.java
@@ -1,8 +1,8 @@
 
-package global-type.resource.cs;
+package global-type-body.resource.cs;
 
 import javax.ws.rs.client.Client;
-import global-type.resource.cs.login.Login;
+import global-type-body.resource.cs.login.Login;
 
 public class Cs {
